Fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -12,7 +12,7 @@ int main(void)
 int i;
 int num1 = 1;
 int num2 = 2;
-int nextnum = num1 + num2;
+long int nextnum = num1 + num2;
 
 printf("%i, %i, %i", num1, num2, nextnum);
 
@@ -21,7 +21,7 @@ for (i = 4; i <= 50; i++)
 num1 = num2;
 num2 = nextnum;
 nextnum = num1 + num2;
- 
+
 printf(", %i", nextnum);
 }
